% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\docType{data}
\name{detector_registry}
\alias{detector_registry}
\alias{.detector_registry}
\title{Registry-based Detector System}
\format{
An object of class \code{environment} of length 2.
}
\usage{
.detector_registry
}
\description{
This section of the package manages a registry for various data leakage detectors. Detectors are stored in the \code{.detector_registry} environment and are accessible by name. The system allows for easy registration of detectors, providing their descriptions and registration times. Detectors can be queried by name or listed.
}
\keyword{internal}
