% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{empty_snapshot_info}
\alias{empty_snapshot_info}
\title{Helper function to return an empty snapshot info dataframe}
\usage{
empty_snapshot_info()
}
\value{
Empty data.frame with correct structure
}
\description{
Helper function to return an empty snapshot info dataframe
}
\keyword{internal}
