% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{export_data_internal}
\alias{export_data_internal}
\title{Export data with consistent messaging}
\usage{
export_data_internal(data, file_path, format, verbose, ...)
}
\arguments{
\item{data}{Data.frame to export}

\item{file_path}{Output file path}

\item{format}{Output format}

\item{verbose}{Whether to show messages}

\item{...}{TODO: Add description}
}
\value{
Path to exported file
}
\description{
Export data with consistent messaging
}
\keyword{internal}
