% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{generate_diagnostic_plots}
\alias{generate_diagnostic_plots}
\title{Generate diagnostic plots for a leakr_report}
\usage{
generate_diagnostic_plots(report)
}
\arguments{
\item{report}{TODO: Document
Generate diagnostic plots for a leakr_report}
}
\value{
A named list of ggplot objects (currently empty stub)
}
\description{
Generate diagnostic plots for a leakr_report
}
\keyword{internal}
