% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_issues_section}
\alias{generate_issues_section}
\title{Generate detailed issues section with output formatting and truncation.}
\usage{
generate_issues_section(report, format)
}
\arguments{
\item{report}{TODO: Document}

\item{format}{TODO: Document}
}
\value{
Formatted issues section string.
}
\description{
Generate detailed issues section with output formatting and truncation.

FIX ME
}
\keyword{internal}
