% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_recommendations}
\alias{generate_recommendations}
\title{Generate actionable recommendations based on report findings.}
\usage{
generate_recommendations(report)
}
\arguments{
\item{report}{A \code{leakr_report} object containing the summary of issues and metadata.}
}
\value{
A character vector of recommendations.
}
\description{
This function generates actionable recommendations based on the findings in a \code{leakr_report} object.
}
\examples{
\dontrun{
recommendations <- generate_recommendations(report)
}

}
\keyword{internal}
