% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_recommendations_section}
\alias{generate_recommendations_section}
\title{Format recommendations for output.}
\usage{
generate_recommendations_section(report, format)
}
\arguments{
\item{report}{A leakr_report object.}

\item{format}{TODO: Add description}
}
\value{
Formatted recommendation section string.
}
\description{
Format recommendations for output.
}
\keyword{internal}
