% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_csv}
\alias{import_csv}
\title{Import CSV files with robust parsing}
\usage{
import_csv(file_path, encoding, verbose, ...)
}
\arguments{
\item{file_path}{Path to CSV file}

\item{encoding}{Character encoding}

\item{verbose}{Whether to show messages}

\item{...}{TODO: Add description}
}
\value{
data.frame
}
\description{
Import CSV files with robust parsing
}
\keyword{internal}
