% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_excel}
\alias{import_excel}
\title{Import Excel files with enhanced sheet support}
\usage{
import_excel(file_path, sheet, verbose, ...)
}
\arguments{
\item{file_path}{Path to Excel file}

\item{sheet}{Sheet name or number}

\item{verbose}{Whether to show messages}

\item{...}{TODO: Add description}
}
\value{
data.frame
}
\description{
Import Excel files with enhanced sheet support
}
\keyword{internal}
