% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_json}
\alias{import_json}
\title{Import JSON files with better structure handling}
\usage{
import_json(file_path, verbose = FALSE, ...)
}
\arguments{
\item{file_path}{Path to the JSON file.}

\item{verbose}{Logical flag indicating whether to show progress messages (default is FALSE).}

\item{...}{Additional arguments passed to \code{jsonlite::fromJSON()}.}
}
\value{
A data.frame with the content from the JSON file, flattened.
}
\description{
Import and process JSON files, converting them into a standardized data.frame.
}
\keyword{internal}
