% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_parquet}
\alias{import_parquet}
\title{Import Parquet files}
\usage{
import_parquet(file_path, verbose = FALSE, ...)
}
\arguments{
\item{file_path}{Path to the Parquet file.}

\item{verbose}{Logical flag indicating whether to show progress messages (default is FALSE).}

\item{...}{Additional arguments passed to \code{arrow::read_parquet()}.}
}
\value{
A data.frame with the content from the Parquet file.
}
\description{
Import and process Parquet files into a standardized data.frame.
}
\keyword{internal}
