% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_rds}
\alias{import_rds}
\title{Import RDS files with validation}
\usage{
import_rds(file_path, verbose, ...)
}
\arguments{
\item{file_path}{Path to RDS file}

\item{verbose}{Whether to show messages}

\item{...}{TODO: Add description}
}
\value{
data.frame
}
\description{
Import RDS files with validation
}
\keyword{internal}
