% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{import_tsv}
\alias{import_tsv}
\title{Import TSV files with robust parsing}
\usage{
import_tsv(file_path, encoding, verbose, ...)
}
\arguments{
\item{file_path}{Path to TSV file}

\item{encoding}{Character encoding}

\item{verbose}{Whether to show messages}

\item{...}{TODO: Add description}
}
\value{
data.frame
}
\description{
Import TSV files with robust parsing
}
\keyword{internal}
