% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{leakr_from_tidymodels}
\alias{leakr_from_tidymodels}
\title{Convert tidymodels workflow to standard format}
\usage{
leakr_from_tidymodels(workflow, data)
}
\arguments{
\item{workflow}{tidymodels workflow object}

\item{data}{Original training data}
}
\value{
List with data and metadata
}
\description{
Extract data from tidymodels workflows for leakage analysis.
}
