% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{leakr_plot}
\alias{leakr_plot}
\title{Plot leakage detection results}
\usage{
leakr_plot(x, ...)
}
\arguments{
\item{x}{Results from leakr_audit}

\item{...}{TODO: Add description
Plot leakage detection results}
}
\value{
A ggplot object
}
\description{
Plot leakage detection results
}
