% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{leakr_quick_import}
\alias{leakr_quick_import}
\title{Fast import with default preprocessing}
\usage{
leakr_quick_import(source, ...)
}
\arguments{
\item{source}{File path or data.frame}

\item{...}{TODO: Add description}
}
\value{
Standardised data.frame
}
\description{
Minimal quick import for typical user workflows. Uses leakr_import internally.
}
