% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{list_registered_detectors}
\alias{list_registered_detectors}
\title{List Registered Detectors}
\usage{
list_registered_detectors()
}
\value{
A character vector containing the names of all registered detectors.
}
\description{
Returns the names of all detectors currently registered in the system. This is useful for checking which detectors are available.
}
\examples{
list_registered_detectors()

}
