% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_temporal_detector}
\alias{new_temporal_detector}
\title{Create a new temporal detector}
\usage{
new_temporal_detector(time_col, lookahead_window = 1)
}
\arguments{
\item{time_col}{Character. Name of the time column}

\item{lookahead_window}{Numeric. Lookahead window size (default 1)
Create a new temporal detector}
}
\value{
A temporal_detector object

A temporal_detector object
}
\description{
Create a new temporal detector
}
