% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_train_test_detector}
\alias{new_train_test_detector}
\title{Create a new train-test detector}
\usage{
new_train_test_detector(threshold = 0.1)
}
\arguments{
\item{threshold}{TODO: Document
Create a new train-test detector}
}
\value{
A train_test_detector object
}
\description{
Create a new train-test detector
}
