% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot.detector_result}
\alias{plot.detector_result}
\title{Plot a detector_result object}
\usage{
\method{plot}{detector_result}(x, palette = NULL, ...)
}
\arguments{
\item{x}{TODO: Document}

\item{palette}{TODO: Document}

\item{...}{TODO: Document}
}
\value{
A ggplot object, invisibly. Printed if interactive

A ggplot object, invisibly. Printed if interactive
}
\description{
Plot a detector_result object

Plot a detector_result object
}
