% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot.udld_report}
\alias{plot.udld_report}
\title{Plot a udld_report object}
\usage{
\method{plot}{udld_report}(x, palette = NULL, ...)
}
\arguments{
\item{x}{A \code{udld_report} object. This object contains the detectors and their associated issues.}

\item{palette}{Optional. A \code{ggplot2} discrete palette for coloring the bars based on severity.}

\item{...}{Additional arguments passed to \code{ggplot2} functions or other methods. These are typically
used for customizing the plot further.}
}
\value{
A \code{ggplot} object, invisibly. The plot is printed if the session is interactive.
}
\description{
This function generates a bar plot of leakage issues detected by different detectors.
The plot displays the count of issues by severity level for each detector in a \code{udld_report} object.
}
