% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{print.leakr_report}
\alias{print.leakr_report}
\title{Print method for leakr_report}
\usage{
\method{print}{leakr_report}(x, ...)
}
\arguments{
\item{x}{leakr_report object}

\item{...}{TODO: Add description}
}
\description{
Print method for leakr_report
}
\keyword{internal}
