% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-detector.R
\name{register_detector}
\alias{register_detector}
\title{Register a new detector}
\usage{
register_detector(name, fun, description = "")
}
\arguments{
\item{name}{Name of the detector}

\item{fun}{TODO: Add description}

\item{description}{TODO: Add description}
}
\value{
Invisibly returns registration status
}
\description{
Register a new data leakage detector function
}
