% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{stratified_sample}
\alias{stratified_sample}
\title{Stratified sampling helper}
\usage{
stratified_sample(target_vec, n_sample)
}
\arguments{
\item{target_vec}{A vector representing the target variable used for stratification.
The function will sample from each class (level) proportionally.}

\item{n_sample}{The total number of samples to draw.}
}
\value{
A vector of indices representing the sampled observations.
}
\description{
This function performs stratified sampling based on the provided target vector.
The sampling is done proportionally to the distribution of values in the target vector.
}
\examples{
\dontrun{
target <- sample(c("class1", "class2", "class3"), size = 100, replace = TRUE)
n_sample <- 30
sampled_indices <- stratified_sample(target, n_sample)
}

}
\keyword{internal}
